Rework comm handling
Only CommHandler now appears in Interpreter. The former CommManager
is renamed to CommTargetManager, and is really only for internal use.
